refactor: move the rpc/spy code to internal/tests pkg
- Updated test files to use the new import
